+++ check_files/reify_anomaly_score.py
@@ -5,12 +5,11 @@ source1 = api.create_source("iris.csv")
 api.ok(source1)
 
 dataset1 = api.create_dataset(source1, \
-    {'name': u'iris dataset'})
+    {'name': u'iris'})
 api.ok(dataset1)
 
 anomaly1 = api.create_anomaly(dataset1, \
-    {'anomaly_seed': u'2c249dda00fbf54ab4cdd850532a584f286af5b6',
-     'name': u"iris dataset's anomaly detector"})
+    {'anomaly_seed': u'2c249dda00fbf54ab4cdd850532a584f286af5b6', 'name': u'iris'})
 api.ok(anomaly1)
 
 anomalyscore1 = api.create_anomaly_score(anomaly1, \