+++ check_files/reify_batch_anomaly_score_dataset_py3.py
@@ -5,18 +5,15 @@
     api.ok(source1)
 
     dataset1 = api.create_dataset(source1, \
-        {'name': 'iris dataset'})
+        {'name': 'iris'})
     api.ok(dataset1)
 
     anomaly1 = api.create_anomaly(dataset1, \
-        {'anomaly_seed': '2c249dda00fbf54ab4cdd850532a584f286af5b6',
-         'name': "iris dataset's anomaly detector"})
+        {'anomaly_seed': '2c249dda00fbf54ab4cdd850532a584f286af5b6', 'name': 'iris'})
     api.ok(anomaly1)
 
     batchanomalyscore1 = api.create_batch_anomaly_score(anomaly1, dataset1, \
-        {'name': "Batch Anomaly Score of iris dataset's anomaly detector with iris "
-                 'dataset',
-         'output_dataset': True})
+        {'name': 'iris using iris dataset', 'output_dataset': True})
     api.ok(batchanomalyscore1)
 
     dataset2 = api.get_dataset(batchanomalyscore1['object']['output_dataset_resource'])