+++ check_files/reify_batch_centroid.py
@@ -5,11 +5,11 @@ source1 = api.create_source("iris.csv")
 api.ok(source1)
 
 dataset1 = api.create_dataset(source1, \
-    {'name': u'iris dataset'})
+    {'name': u'iris'})
 api.ok(dataset1)
 
 cluster1 = api.create_cluster(dataset1, \
-    {'name': u"iris dataset's cluster"})
+    {'name': u'iris'})
 api.ok(cluster1)
 
 batchcentroid1 = api.create_batch_centroid(cluster1, dataset1, \