+++ check_files/reify_batch_centroid_dataset.py
@@ -5,16 +5,15 @@ source1 = api.create_source("iris.csv")
 api.ok(source1)
 
 dataset1 = api.create_dataset(source1, \
-    {'name': u'iris dataset'})
+    {'name': u'iris'})
 api.ok(dataset1)
 
 cluster1 = api.create_cluster(dataset1, \
-    {'name': u"iris dataset's cluster"})
+    {'name': u'iris'})
 api.ok(cluster1)
 
 batchcentroid1 = api.create_batch_centroid(cluster1, dataset1, \
-    {'name': u"Batch Centroid of iris dataset's cluster with iris dataset",
-     'output_dataset': True})
+    {'name': u'iris using iris dataset', 'output_dataset': True})
 api.ok(batchcentroid1)
 
 dataset2 = api.get_dataset(batchcentroid1['object']['output_dataset_resource'])