+++ check_files/reify_dataset_dataset.py
@@ -5,7 +5,7 @@ source1 = api.create_source("iris.csv")
 api.ok(source1)
 
 dataset1 = api.create_dataset(source1, \
-    {'name': u'iris dataset'})
+    {'name': u'iris'})
 api.ok(dataset1)
 
 dataset2 = api.create_dataset(dataset1, \