+++ check_files/reify_dataset_dataset_py3.py
@@ -5,7 +5,7 @@
     api.ok(source1)
 
     dataset1 = api.create_dataset(source1, \
-        {'name': 'iris dataset'})
+        {'name': 'iris'})
     api.ok(dataset1)
 
     dataset2 = api.create_dataset(dataset1, \