+++ check_files/reify_evaluation.py
@@ -5,11 +5,11 @@ source1 = api.create_source("iris.csv")
 api.ok(source1)
 
 dataset1 = api.create_dataset(source1, \
-    {'name': u'iris dataset'})
+    {'name': u'iris'})
 api.ok(dataset1)
 
 model1 = api.create_model(dataset1, \
-    {'name': u"iris dataset's model"})
+    {'name': u'iris'})
 api.ok(model1)
 
 evaluation1 = api.create_evaluation(model1, dataset1, \