+++ check_files/reify_evaluation_split.py
@@ -5,24 +5,19 @@ source1 = api.create_source("iris.csv")
 api.ok(source1)
 
 dataset1 = api.create_dataset(source1, \
-    {'name': u'iris dataset'})
+    {'name': u'iris'})
 api.ok(dataset1)
 
 dataset2 = api.create_dataset(dataset1, \
-    {'name': u'iris dataset - sample (70.00%)',
-     'sample_rate': 0.7,
-     'seed': u'BigML'})
+    {'name': u'iris', 'sample_rate': 0.7, 'seed': u'BigML'})
 api.ok(dataset2)
 
 dataset3 = api.create_dataset(dataset1, \
-    {'name': u'iris dataset - sample (30.00%)',
-     'out_of_bag': True,
-     'sample_rate': 0.7,
-     'seed': u'BigML'})
+    {'name': u'iris', 'out_of_bag': True, 'sample_rate': 0.7, 'seed': u'BigML'})
 api.ok(dataset3)
 
 model1 = api.create_model(dataset2, \
-    {'name': u"iris dataset - sample (70.00%)'s model"})
+    {'name': u'iris'})
 api.ok(model1)
 
 evaluation1 = api.create_evaluation(model1, dataset3, \