+++ check_files/reify_evaluation_split_py3.py
@@ -5,22 +5,19 @@
     api.ok(source1)
 
     dataset1 = api.create_dataset(source1, \
-        {'name': 'iris dataset'})
+        {'name': 'iris'})
     api.ok(dataset1)
 
     dataset2 = api.create_dataset(dataset1, \
-        {'name': 'iris dataset - sample (70.00%)', 'sample_rate': 0.7, 'seed': 'BigML'})
+        {'name': 'iris', 'sample_rate': 0.7, 'seed': 'BigML'})
     api.ok(dataset2)
 
     dataset3 = api.create_dataset(dataset1, \
-        {'name': 'iris dataset - sample (30.00%)',
-         'out_of_bag': True,
-         'sample_rate': 0.7,
-         'seed': 'BigML'})
+        {'name': 'iris', 'out_of_bag': True, 'sample_rate': 0.7, 'seed': 'BigML'})
     api.ok(dataset3)
 
     model1 = api.create_model(dataset2, \
-        {'name': "iris dataset - sample (70.00%)'s model"})
+        {'name': 'iris'})
     api.ok(model1)
 
     evaluation1 = api.create_evaluation(model1, dataset3, \